requirements.txt
pyTelegramBotAPI==4.2.0
googlesearch-python==1.0.1

